[ProLib] Inheritance update
